split tests to many files
