+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Tests/WithTextShape_SceneMetricsCounterShould.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e4accca85ea04aa6a74360dfb7578c36
+timeCreated: 1644097837
